refactor: extract structs into separate files
